+++ vaapi/vaapiutils.cpp
@@ -1,5 +1,5 @@
 /*
- *  vaapiimageutils.h - vaapi image utils
+ *  vaapiutils.cpp - vaapi image utils
  *
  *  Copyright (C) 2011-2014 Intel Corporation
  *    Author: Xu Guangxin<guangxin.xu@intel.com>
@@ -20,13 +20,31 @@
  *  Boston, MA 02110-1301 USA
  */
 
-#include <stdint.h>
-#include <va/va.h>
+#ifdef HAVE_CONFIG_H
+#include "config.h"
+#endif
 
-namespace YamiMediaCodec{
+#include "vaapiutils.h"
 
-uint8_t* mapSurfaceToImage(VADisplay display, intptr_t surface, VAImage& image);
+namespace YamiMediaCodec {
 
-void unmapImage(VADisplay display, const VAImage& image);
+uint8_t* mapSurfaceToImage(VADisplay display, intptr_t surface, VAImage& image)
+{
+    uint8_t* p = NULL;
+    VAStatus status = vaDeriveImage(display, (VASurfaceID)surface, &image);
+    if (!checkVaapiStatus(status, "vaDeriveImage"))
+        return NULL;
+    status = vaMapBuffer(display, image.buf, (void**)&p);
+    if (!checkVaapiStatus(status, "vaMapBuffer")) {
+        checkVaapiStatus(vaDestroyImage(display, image.image_id), "vaDestroyImage");
+        return NULL;
+    }
+    return p;
+}
 
-} //namespace YamiMediaCodec
+void unmapImage(VADisplay display, const VAImage& image)
+{
+    checkVaapiStatus(vaUnmapBuffer(display, image.buf), "vaUnmapBuffer");
+    checkVaapiStatus(vaDestroyImage(display, image.image_id), "vaDestroyImage");
+}
+}
