+++ vaapi/vaapiutils.h
@@ -25,6 +25,12 @@
 #include "common/log.h"
 #include <va/va.h>
 
+namespace YamiMediaCodec {
+
+uint8_t* mapSurfaceToImage(VADisplay display, intptr_t surface, VAImage& image);
+
+void unmapImage(VADisplay display, const VAImage& image);
+
 #define checkVaapiStatus(status, prompt)                     \
     (                                                        \
         {                                                    \
@@ -34,5 +40,6 @@
                 ERROR("%s: %s", prompt, vaErrorStr(status)); \
             ret;                                             \
         })
+}
 
 #endif //vaapiutils_h
